svelte/conditional_prompt/demo/requirements.txt
gradio_conditional_prompt